+++ example/stairLift.py
@@ -2,6 +2,7 @@ import cadquery as cq
 from cqspoolterrain import StairLift
 
 bp_stairs = StairLift()
+bp_stairs.bp_stairs.render_hollow = False
 bp_stairs.make()
 stairs = bp_stairs.build()
 #show_object(stairs)